% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{catchment}
\alias{catchment}
\title{Simulated catchment area data}
\format{
A data frame with 16 rows and 6 variables
}
\usage{
data(catchment)
}
\description{
A data frame simulated from code catchment_sim(16, 0.5, 1.5, 5)
}
\details{
\itemize{
\item catchID, ID associated with each catchment
\item num.schools, Number of schools in each catchment
\item xStart, Start x coordinate of catchment grid
\item xEnd, End x coordinate of catchment grid
\item yStart, Start y coordinate of catchment grid
\item yEnd, End y coordinate of catchment grid
}
}
