% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elementary_pop.R
\name{elementary_pop}
\alias{elementary_pop}
\title{Create Elementary Schools population size}
\usage{
elementary_pop(df, alpha, beta)
}
\arguments{
\item{df}{output data frame from catchment_sim function}

\item{alpha}{shape parameter of gamma distribution}

\item{beta}{rate parameter of gamma distribution}
}
\value{
a data frame of simulated elementary school population
}
\description{
Function to simulation elementary school size and then assigned to catchment areas.
}
\examples{
catch_df <- catchment_sim(16, 4.313320, 3.026894, 20)
elementary_df <- elementary_pop(catch_df, 5.27426341, 0.01427793)


}
