% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{epidemic}
\alias{epidemic}
\title{Simulated epidemics data}
\format{
A data frame with 5,599 rows and 13 variables
}
\usage{
data(epidemic)
}
\description{
A list of 10 lists simulated from code simepi(individuals, b=3, sus=.0019, spark=0, num_inf = 2, rep = 10)
Simulated through epidata() from EpiILM package. Each data frame follows the format below:
}
\details{
\itemize{
\item XYcoordinates, A data frame of x and y coordinates associated with each household
\item inftime, Day in which each individual got infected
\item remtime, Day in which each individual is removed from infection
\item start, Day epidemic started
}
}
