% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_alarm_metrics.R
\name{eval_alarm_metrics}
\alias{eval_alarm_metrics}
\title{Evaluate Alarm Metrics for Epidemic Models}
\usage{
eval_alarm_metrics(
  data,
  ScYr = c(2:10),
  maxlag = 15,
  yr.weights = c(1:9)/sum(c(1:9)),
  thres = seq(0.1, 0.6, by = 0.05)
)
}
\arguments{
\item{ScYr}{A vector of school years to be used for model evaluation (default: 2:10).}

\item{maxlag}{The maximum number of lags to consider (default: 15).}

\item{yr.weights}{Weights for calculating WFATQ and WAATQ metrics (default: (1:9)/sum(1:9)).}

\item{thres}{A vector of threshold values to evaluate (default: seq(0.1, 0.6, by = 0.05)).}

\item{lagdata}{A data frame containing the epidemic data with lagged variables,
typically output from the compile_epi function.}
}
\value{
A list containing:
\item{FAR}{Matrix of False Alarm Rates for each lag and threshold}
\item{ADD}{Matrix of Added Days Delayed for each lag and threshold}
\item{AATQ}{Matrix of Average Alarm Time Quality for each lag and threshold}
\item{FATQ}{Matrix of First Alarm Time Quality for each lag and threshold}
\item{WAATQ}{Matrix of Weighted Average Alarm Time Quality for each lag and threshold}
\item{WFATQ}{Matrix of Weighted First Alarm Time Quality for each lag and threshold}
\item{best.AATQ}{Best model according to AATQ}
\item{best.FATQ}{Best model according to FATQ}
\item{best.FAR}{Best model according to FAR}
\item{best.ADD}{Best model according to ADD}
\item{best.WFATQ}{Best model according to WFATQ}
\item{best.WAATQ}{Best model according to WAATQ}
}
\description{
This function calculates various performance metrics for epidemic alarm systems
across different lags and thresholds. It evaluates False Alarm Rate (FAR),
Added Days Delayed (ADD), Average Alarm Time Quality (AATQ), First Alarm Time
Quality (FATQ), and their weighted versions (WAATQ, WFATQ).
}
\examples{
# Generate simulated epidemic data
epidemic <- list(year1 = list(new_inf = rnorm(300), reported_cases = rnorm(300)),
                 year2 = list(new_inf = rnorm(300), reported_cases = rnorm(300)))
individual_data <- data.frame(elem_child_ind = sample(0:1, 1000, replace = TRUE),
                              schoolID = sample(1:10, 1000, replace = TRUE))

compiled_data <- compile_epi(epidemic, individual_data)

# Evaluate alarm metrics
alarm_metrics <- eval_alarm_metrics(compiled_data, thres = seq(0.1,0.6,by = 0.05),
                                  ScYr = c(2:10), yr.weights = c(1:9)/sum(c(1:9)))

# View the best model according to AATQ
print(alarm_metrics$best.AATQ)

}
\seealso{
\code{\link{ssir}}, \code{\link{compile_epi}}
}
