% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_metrics.R
\name{eval_metrics}
\alias{eval_metrics}
\title{Calculate evaluation metrics for lag, threshold, and yearly model}
\usage{
eval_metrics(
  type = "r",
  lagdata,
  mod_resp,
  ScYr,
  yr.weights,
  lags = seq.int(1, 15),
  thres = seq(0.1, 0.6, by = 0.05)
)
}
\arguments{
\item{type}{'c' for catchment area models, 'r' for region-wide models. Default is region-wide}

\item{lagdata}{lagged dataframe used for modelling}

\item{mod_resp}{output from "log_reg" function}

\item{ScYr}{vector of all school years to be used for model evaluation (range of natural numbers)}

\item{yr.weights}{weights for WFATQ and WAATQ metrics}

\item{lags}{lags evaluated in model}

\item{thres}{thresholds evaluated}
}
\value{
List of matrices for FAR, ADD, AATQ, FATQ, WAATQ, WFATQ
}
\description{
Calculate evaluation metrics for lag, threshold, and yearly model
}
\examples{
\dontrun{I need to update this}

}
