% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{house_children}
\alias{house_children}
\title{Simulated households with children data}
\format{
A data frame with 491 rows and 12 variables
}
\usage{
data(house_children)
}
\description{
A data frame simulated from code the subpop_children(elementary)
Proportion of parents as a couple = 0.7668901
proportion of coupled parents with 1, 2, 3+ children = 0.3634045, 0.4329440, 0.2036515
proportion of single parents with 1, 2, 3+ children = 0.5857832, 0.3071523, 0.1070645
proportion of children that are of elementary school age = 0.4976825
}
\details{
\describe{
\item {schoolID}{ID associated with schools in each catchment}
\item {houseID}{ID associated with each household}
\item {num_parent}{Number of parents in each household}
\item {num_child}{Number of children in each household}
\item {num_elem_child}{Number of children who are of elementary school age}
\item {catchID}{ID associated with each catchment}
\item {schoolPop}{Population of students in each school generated from a gamma distribution}
\item {xStart}{Start x coordinate of catchment grid}
\item {xEnd}{End x coordinate of catchment grid}
\item {yStart}{Start y coordinate of catchment grid}
\item {yEnd}{End y coordinate of catchment grid}
\item {num_people}{Total number of people in each household}
}
}
