% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{house_nochildren}
\alias{house_nochildren}
\title{Simulated households with no children data}
\format{
A data frame with 656 rows and 8 variables
}
\usage{
data(house_nochildren)
}
\description{
A data frame simulated from code the subpop_noChildren(house_children, elementary)
proportion of households with 1, 2, 3, 4, 5+ members = 0.23246269, 0.34281716, 0.16091418, 0.16427239, 0.09953358
proportion of households with children = 0.4277052
}
\details{
\describe{
\item {catchID}{ID associated with each catchment}
\item {houseID}{ID associated with each household}
\item {num_people}{Total number of people in each household}
\item {schoolPop}{Population of students in each school generated from a gamma distribution}
\item {xStart}{Start x coordinate of catchment grid}
\item {xEnd}{End x coordinate of catchment grid}
\item {yStart}{Start y coordinate of catchment grid}
\item {yEnd}{End y coordinate of catchment grid}
}
}
