% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_reg.R
\name{log_reg}
\alias{log_reg}
\title{Logistic regression model fitting function for simulated data}
\usage{
log_reg(lagdata, maxlag = 15)
}
\arguments{
\item{lagdata}{Data frame with lagged absenteeism data}

\item{maxlag}{maximum lagged day to be assessed}
}
\value{
list of lists of linear models (mod) and model responses (resp)
mod is a list of list of yearly models created using all data temporarily preceded that year (i.e, list length is equal to number of years)

resp contains responses for each year. Length of list is the number of models fitted (number of lags).
}
\description{
Logistic regression model fitting function for simulated data
}
\examples{
 \dontrun{I need to update this}


}
