% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_data.R
\name{model_data}
\alias{model_data}
\title{Title}
\usage{
model_data(epi.series, individual.data)
}
\arguments{
\item{epi.series}{simulated epidemic data frame}

\item{individual.data}{individuals data frame}
}
\value{
returns simulated data frame
}
\description{
Title
}
\examples{
#' \dontrun{I need to update this}

}
