% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{region}
\alias{region}
\title{Simulated region data}
\format{
A data frame with 3,000 rows and 29 variables
}
\usage{
data(region)
}
\description{
A data frame simulated from code model_data(epidemic, individuals, type = "r")
region is one of two data frames contained within output of model_data() function.
region contains information about confirmed cases for the region.
}
\details{
\itemize{
\item Date, Day (1-300)
\item ScYr, School year
\item Actual.case, Total number of true cases (unconfirmed)
\item case.elem, Number of true cases that are elementary school children
\item pct.absent, Percent of students absent in total school population
\item absent, Total number of students absent
\item absent.sick, Total number of students absent due to illness
\item Case.No, Number of lab confirmed cases
\item Case, Binary variable to indicate whether there has been a lab confirmed case for the day
\item sinterm, Seasonal sine term
\item costerm, Seasonal cosine term
\item window, 14 day true alarm window for ADD and FAR calculations (binary variable)
\item ref.date, Reference date (first day where there are two lab confirmed cases in a 7 day period)
\item lag0, lag1, ..., lag15, Lagged absenteeism values
}
}
