% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simepi.R
\name{sim12epiNew}
\alias{sim12epiNew}
\title{Simulate epidemic}
\usage{
sim12epiNew(df, b, sus, spar, num_inf)
}
\arguments{
\item{df}{dataframe}

\item{b}{parameter}

\item{sus}{parameter}

\item{spar}{parameter}

\item{num_inf}{parameter}
}
\value{
list of simulated epidemic
}
\description{
simulate epidemic
}
\examples{
\dontrun{whaever }
}
