% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subpop_children.R
\name{subpop_children}
\alias{subpop_children}
\title{Simulate households with children}
\usage{
subpop_children(df, n = 5)
}
\arguments{
\item{df}{simulated output data frame from elementary_pop function}

\item{n}{population multiplier, default value = 5,}
}
\value{
simulated population of households with children
}
\description{
Simulation of households with children via random uniform distribution. Number of observations is multiplied by a default value of five.
}
\details{
This is an interactive function where user has to enter eight population proportion values as prompted.
To automate entering these prompts, establish a file connection where a vector of the eight population proportion values are written to an open file. Refer to examples section.
}
\examples{
# simulate catchment area
 catch_df <- catchment_sim(8, 4.30, 3.6, 5)

# simulate elementary schools for each area
 elementary_df <- elementary_pop(catch_df, 5.7, 0.014)

# Establish a file connection for population proportion values when prompted by subpop_children()
 f <- file()
 lines <- c(0.7668901,0.3634045, 0.4329440, 0.2036515,0.5857832, 0.3071523, 0.1070645,0.4976825)
 ans <- paste(lines, collapse = "\n")
 write(ans, f)
 options("usr_con" = f)#'

# simulate household with children and assign them to elementary school
 house_children <- subpop_children(elementary_df)

 # close the file
 close(f)

 # reset connection option
 options("usr_con" = stdin())


}
