% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subpop_noChildren.R
\name{subpop_noChildren}
\alias{subpop_noChildren}
\title{Simulate households without children}
\usage{
subpop_noChildren(df, df2)
}
\arguments{
\item{df}{simulated output data frame from subpop_children function}

\item{df2}{simulated output data frame from elementary_pop function}
}
\value{
simulated population of households without children
}
\description{
Simulation of households without children via random uniform distribution using data frames from elementary_pop() and house_children() functions
}
\examples{
\dontrun{
#simulate catchment area
catch_df <- catchment_sim(16, 4.313320, 3.026894, 20)

#simulate elementary schools for each area
elementary_df <- elementary_pop(catch_df, 5.27426341, 0.01427793)

# simulate household with children and assign them to elementary school
house_children <- subpop_children(elementary_df)

# simulate household with no children and assign them to elementary school
house_noChild <- subpop_noChildren(house_children, elementary_df)
}

}
