% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.ssir_epidemic.R
\name{summary.ssir_epidemic}
\alias{summary.ssir_epidemic}
\title{S3 Summary method for epidemic Simulation}
\usage{
\method{summary}{ssir_epidemic}(object, ...)
}
\arguments{
\item{object}{An object of class "ssir_epidemic", typically the result of
calling ssir() with rep = 1 or NULL.}

\item{...}{Additional arguments affecting the summary produced.}
}
\value{
No return value, called for side effects.
}
\description{
This function provides a summary of SSIR epidemic simulation.
It calculates and displays the  statistics.
}
\examples{
# Run multiple simulations
result <- ssir(N = 10000, T = 300, alpha = 0.3, inf_period = 4,
                     inf_init = 32, report = 0.02, lag = 7)

# Display summary
summary(result)

}
