% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catchment_sim.R
\name{catchment_sim}
\alias{catchment_sim}
\title{Simulating catchment areas}
\usage{
catchment_sim(n, alpha, beta, area)
}
\arguments{
\item{n}{number of catchments to be simulated}

\item{alpha}{shape parameter of gamma distribution for number of schools to be simulated}

\item{beta}{rate parameter of gamma distribution for number of schools to be simulated}

\item{area}{square dimension for catchment (if area = 20, then each catchment dimensions will be 20 x 20)}
}
\value{
data frame of n observations
}
\description{
Function to simulate specified catchments of square area (a x a).
Catchment areas are simulated via gamma distribution.
}
\examples{
catch_df <- catchment_sim(16, 4.1, 2.7, 20)


}
