% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elementary_pop.R
\name{elementary_pop}
\alias{elementary_pop}
\title{Create Elementary Schools population size}
\usage{
elementary_pop(df, alpha, beta)
}
\arguments{
\item{df}{output data frame from catchment_sim function}

\item{alpha}{shape parameter of gamma distribution to simulate
school population}

\item{beta}{rate parameter of gamma distribution to simulate
school population}
}
\value{
a data frame of simulated elementary school population
}
\description{
Function to simulate elementary school size and assigns it to catchments
}
\examples{
catch_df <- catchment_sim(16, 3.5, 2.8, 20)
elementary_df <- elementary_pop(catch_df, 5.1, 0.015)


}
