% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_data.R
\name{model_data}
\alias{model_data}
\title{Simulate laboratory confirmed cases and school absenteeism data}
\usage{
model_data(epi.series, individual.data, type = "r")
}
\arguments{
\item{epi.series}{simulated epidemic data frame
(output from simepi() function)}

\item{individual.data}{individuals data frame
(output from simulate_households() function)}

\item{type}{type of analysis (regional or catchment)}
}
\value{
returns data frame
}
\description{
Simulates lab confirmed cases and absenteeism information using epidemics
and individuals data. Simulation can be performed at the regional
or catchment level.
}
\examples{
#Simulate catchment data
catch_df <- catchment_sim(4, 4.12, 3.01, 5)

#simulate elementary schools for each area
elementary_df <- elementary_pop(catch_df, 4.8, 0.015)

# Enters values for prompts from subpop_children() function
f <- file()
lines <- c(0.7668901,0.3634045, 0.4329440, 0.2036515, 0.5857832,
              0.3071523, 0.1070645,0.4976825)

ans <- paste(lines, collapse = "\n")
write(ans, f)
options("usr_con" = f) # set connection option

# simulating households without children
house_children <- subpop_children(elementary_df, n = 2)

# Enters values for prompts from subpop_nochildren() function
lines <- c(0.23246269, 0.34281716, 0.16091418, 0.16427239,
               0.09953358, 0.4277052)

ans <- paste(lines, collapse = "\n")
write(ans, f)

# simulate households and individuals data
house_nochildren <- subpop_noChildren(house_children, elementary_df)

close(f) # close the file
options("usr_con" = stdin()) # reset connection option

# simulate households and individuals data
simulation <- simulate_households(house_children, house_nochildren)

# randomly sampling 1000 rows to reduce simulation times
individuals <- simulation$individual_sim[sample(nrow(
                                          simulation$individual_sim),200),]

# simulate epidemic
epidemic <- simepi(individuals, b=3, sus=.0019, spark=0,
                     num_inf = 2, rep = 1)

# simulate laboratory confirmed cases, and school absenteeism data sets
data <- model_data(epidemic, individuals, type = "r")

}
