% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simepi.R
\name{simepi}
\alias{simepi}
\title{Simulate epidemic}
\usage{
simepi(df, b, sus, spark, num_inf)
}
\arguments{
\item{df}{individuals data frame from household data simulation}

\item{b}{Spatial parameter(s) (>0) or network parameter (s) (>0) if contact network is used.}

\item{sus}{Susceptibility parameter (>0)}

\item{spark}{Sparks parameter (>=0), representing infections unexplained by other parts of the model.}

\item{num_inf}{number of individuals that are infected initially in each catchment}
}
\value{
list of 10 simulated epidemics
}
\description{
This function under Susceptible-Infectious-Removed (SIR) compartment framework. The time period for simulation is 270 days
}
\examples{
\dontrun{I need to update this}

}
