% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_households.R
\name{simulate_households}
\alias{simulate_households}
\title{Simulate total households and individuals data}
\usage{
simulate_households(children_df, noChildren_df)
}
\arguments{
\item{children_df}{data frame output from house_children() function}

\item{noChildren_df}{data frame output from house_noChildren() function}
}
\value{
list of two data frames; simulated individuals data and simulated households data
}
\description{
Creates two simulated populations, households and individuals
}
\examples{
# simulate catchment area
 catch_df <- catchment_sim(8, 4.30, 3.6, 5)

# simulate elementary schools for each area
 elementary_df <- elementary_pop(catch_df, 5.7, 0.014)

# Establish a file connection for population proportion values when prompted by subpop_children()
 f <- file()
 lines <- c(0.77, 0.36, 0.43, 0.21,0.59, 0.31, 0.10, 0.49)
 ans <- paste(lines, collapse = "\n")
 write(ans, f)
 options("usr_con" = f)

# simulate household with children and assign them to elementary school
 house_children <- subpop_children(elementary_df, n = 2)

# Overwriting connection with population proportion values for households without children
 lines <- c(0.23, 0.34, 0.17, 0.16, 0.1, 0.43)
 ans <- paste(lines, collapse = "\n")
 write(ans, f)

# simulate household with no children and assign them to elementary school
 house_nochildren <- subpop_noChildren(house_children, elementary_df)

# close the file
 close(f)
# reset connection option
 options("usr_con" = stdin())

 # simulate households and individuals data
 simulation <- simulate_households(house_children, house_nochildren)

}
