% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_households.R
\name{simulate_households}
\alias{simulate_households}
\title{Simulate total households and individuals data}
\usage{
simulate_households(children_df, noChildren_df)
}
\arguments{
\item{children_df}{data frame output from house_children() function}

\item{noChildren_df}{data frame output from house_noChildren() function}
}
\value{
list of two data frames
}
\description{
Creates two simulated populations, households and individuals
}
\examples{
\dontrun{
#' #simulate catchment area
catch_df <- catchment_sim(16, 4.313320, 3.026894, 20)

#simulate elementary schools for each area
elementary_df <- elementary_pop(catch_df, 5.27426341, 0.01427793)

# simulate household with children and assign them to elementary school
house_children <- subpop_children(elementary_df)

# simulate household with no children and assign them to elementary school
house_noChild <- subpop_noChildren(house_children, elementary_df)

#simulate total households and individuals
simulation <- simulate_households(house_children, house_noChild)
}


}
