% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_households.R
\name{simulate_households}
\alias{simulate_households}
\title{Simulate total households and individuals data}
\usage{
simulate_households(children_df, noChildren_df)
}
\arguments{
\item{children_df}{data frame output from house_children() function}

\item{noChildren_df}{data frame output from house_noChildren() function}
}
\value{
list of two data frames; simulated individuals data
and simulated households data
}
\description{
Creates two simulated populations, households and individuals
}
\examples{
#Simulate catchment data
catch_df <- catchment_sim(4, 1.2, 3.01, 5)

#simulate elementary schools for each area
elementary_df <- elementary_pop(catch_df, 0.5, 0.015)

# Enters values for prompts from subpop_children() function
f <- file()
lines <- c(0.7668901,0.3634045, 0.4329440, 0.2036515,0.5857832,
                0.3071523, 0.1070645,0.4976825)

ans <- paste(lines, collapse = "\n")
write(ans, f)
options("usr_con" = f) # set connection option

# simulating households without children
house_children <- subpop_children(elementary_df, n = 2)

# Enters values for prompts from subpop_nochildren() function
lines <- c(0.23246269, 0.34281716, 0.16091418,
                0.16427239, 0.09953358, 0.4277052)

ans <- paste(lines, collapse = "\n")
write(ans, f)

# simulate households and individuals data
house_nochildren <- subpop_noChildren(house_children, elementary_df)

close(f) # close the file
options("usr_con" = stdin()) # reset connection option

# simulate households and individuals data
simulation <- simulate_households(house_children, house_nochildren)

}
