% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subpop_children.R
\name{subpop_children}
\alias{subpop_children}
\title{Simulate households with children}
\usage{
subpop_children(df, n = 5)
}
\arguments{
\item{df}{simulated output data frame from elementary_pop function}

\item{n}{population multiplier, default value = 5,}
}
\value{
simulated population of households with children
}
\description{
Simulation of households with children via random uniform distribution. Number of observations is multiplied by a default value of five.
}
\examples{
\dontrun{#simulate catchment area
 catch_df <- catchment_sim(16, 4.313320, 3.026894, 20)

 #simulate elementary schools for each area
 elementary_df <- elementary_pop(catch_df, 5.27426341, 0.01427793)

 # simulate household with children and assign them to elementary school
 house_children <- subpop_children(elementary_df)
 }



}
